Fix problem with size taggs

+++ app/src/main/java/com/jnsoft/instatagg/data/database/TaggsDao.kt
@@ -29,8 +29,8 @@ interface TaggsDao {
     suspend fun getTagg(id: Long): Tagg
 
     @Query("UPDATE taggs SET size = size - :size WHERE id = :id")
-    suspend fun decreaseTaggSize(size: Int, id: Long)
+    suspend fun decreaseTaggSize(size: Long, id: Long)
 
     @Query("UPDATE taggs SET size = size + :size WHERE id = :id")
-    suspend fun increaseTaggSize(size: Int, id: Long)
+    suspend fun increaseTaggSize(size: Long, id: Long)
 }
