+++ app/src/main/java/com/jnsoft/instatagg/presentation/viewmodel/FullscreanPhotoViewModel.kt
@@ -14,18 +14,18 @@ class FullscreanPhotoViewModel(
     private val mainRepository: MainRepository
 ) : ViewModel() {
 
-    fun delPhoto(photo: Photo, size: Int){
+    fun delPhoto(photo: Photo, size: Long){
         viewModelScope.launch { mainRepository.delPhoto(photo, size) }
     }
     fun movePhoto(newTaggName: String, newTaggColor: Int, newTaggId: Long, id: Long,
-                  size: Int, oldTaggId: Long){
+                  size: Long, oldTaggId: Long){
         viewModelScope.launch { mainRepository.movePhoto(newTaggName, newTaggColor, newTaggId, id,
                                                          size, oldTaggId) }
     }
     fun getTaggs(): LiveData<List<Tagg>> = liveData(Dispatchers.IO) {
         emit(mainRepository.getTaggs())
     }
-    fun insertPhoto(photo: Photo, size: Int){
+    fun insertPhoto(photo: Photo, size: Long){
         viewModelScope.launch { mainRepository.insertPhoto(photo, size) }
     }
     fun getPhotos(id: Long) = liveData(Dispatchers.IO) {
